# backend/openapi.json
{
  "components": {
    "schemas": {
      "AskRequest": {
        "properties": {
          "choice_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Choice Id"
          },
          "program": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Program"
          },
          "selected_object_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Object Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "text": {
            "default": "",
            "title": "Text",
            "type": "string"
          },
          "viewpoint": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ViewpointRequest"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "AskRequest",
        "type": "object"
      },
      "Body_import_package_v1_imports_post": {
        "properties": {
          "file": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          },
          "inbox_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inbox Name"
          },
          "replace": {
            "default": false,
            "title": "Replace",
            "type": "boolean"
          }
        },
        "title": "Body_import_package_v1_imports_post",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ReprocessRequest": {
        "properties": {
          "stage": {
            "title": "Stage",
            "type": "string"
          }
        },
        "required": [
          "stage"
        ],
        "title": "ReprocessRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ViewpointRequest": {
        "properties": {
          "forward": {
            "maxItems": 3,
            "minItems": 3,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Forward",
            "type": "array"
          },
          "position": {
            "maxItems": 3,
            "minItems": 3,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Position",
            "type": "array"
          }
        },
        "required": [
          "position",
          "forward"
        ],
        "title": "ViewpointRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "S3D",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/imports": {
      "post": {
        "operationId": "import_package_v1_imports_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_package_v1_imports_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Import Package V1 Imports Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Package"
      }
    },
...
    "/v1/scenes/{scene_id}": {
      "delete": {
        "operationId": "delete_scene_v1_scenes__scene_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "delete_package",
            "required": false,
            "schema": {
              "default": false,
              "title": "Delete Package",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Scene V1 Scenes  Scene Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Scene"
      },
      "get": {
        "operationId": "get_scene_v1_scenes__scene_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Scene V1 Scenes  Scene Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scene"
      }
    },
    "/v1/scenes/{scene_id}/ask": {
      "post": {
        "operationId": "ask_v1_scenes__scene_id__ask_post",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask"
      }
    },
    "/v1/scenes/{scene_id}/events": {
      "get": {
        "operationId": "scene_events_v1_scenes__scene_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scene Events"
      }
    },
    "/v1/scenes/{scene_id}/frames/{frame_id}.jpg": {
      "get": {
        "operationId": "frame_v1_scenes__scene_id__frames__frame_id__jpg_get",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "frame_id",
            "required": true,
            "schema": {
              "title": "Frame Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Frame"
      }
    },
    "/v1/scenes/{scene_id}/graph": {
      "get": {
        "operationId": "scene_graph_v1_scenes__scene_id__graph_get",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Scene Graph V1 Scenes  Scene Id  Graph Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scene Graph"
      }
    },
    "/v1/scenes/{scene_id}/mesh": {
      "get": {
        "operationId": "scene_mesh_v1_scenes__scene_id__mesh_get",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scene Mesh"
      }
    },
    "/v1/scenes/{scene_id}/objects": {
      "get": {
        "operationId": "scene_objects_v1_scenes__scene_id__objects_get",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "label",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Label"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Scene Objects V1 Scenes  Scene Id  Objects Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scene Objects"
      }
    },
    "/v1/scenes/{scene_id}/objects/{object_id}/mask": {
      "get": {
        "operationId": "object_mask_v1_scenes__scene_id__objects__object_id__mask_get",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "object_id",
            "required": true,
            "schema": {
              "title": "Object Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Object Mask V1 Scenes  Scene Id  Objects  Object Id  Mask Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Object Mask"
      }
    },
    "/v1/scenes/{scene_id}/reprocess": {
      "post": {
        "operationId": "reprocess_scene_v1_scenes__scene_id__reprocess_post",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReprocessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Reprocess Scene V1 Scenes  Scene Id  Reprocess Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reprocess Scene"
      }
    },
    "/v1/scenes/{scene_id}/structures": {
      "get": {
        "operationId": "scene_structures_v1_scenes__scene_id__structures_get",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Scene Structures V1 Scenes  Scene Id  Structures Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scene Structures"
      }
    }
  }
}